Day 4: Converted to longs, made it flexible, added test for 100x100 cards

+++ src/AoC2021.Code/Day4.cs
@@ -5,14 +5,14 @@ namespace AoC2021.Core
 {
 	public static class Day4
 	{
-		public static int Solve1(Day4PuzzleInput input)
+		public static long Solve1(Day4PuzzleInput input)
 		{
 			var (winningCard, winningNumber) = input.PlayBingo();
 			return winningCard.GetSumOfUnmarkedNumbers() * winningNumber;
 		}
 
 		
-		public static int Solve2(Day4PuzzleInput input)
+		public static long Solve2(Day4PuzzleInput input)
 		{
 			var (losingCard, winningNumber) = input.PlayBingoToLose();
 			return losingCard.GetSumOfUnmarkedNumbers() * winningNumber;
